Add edit functionality, make dark mode a bit better
Dark mode is now based off the system preference, rather than the `dark`
class on the `:root`.

The colours could probably still be improved and there are lots of
variables defined that I'm probably not using that were just part of the
starter template for this codebase. I should probably clean these up at
some point.