Remove the hover styles from the table
Of all the sticky hover/focus issues in this app, this was the worst
offender as it looked like you couldn't deselect a row by clicking on it
again as the sticky hover style made it still look selected.

I don't really understand why this is happening, given the hover styles
look like they're wrapped in a media query to prevent this, but removing
is the simplest option for now.

There is still the pointer cursor to show it's clickable.