Use fatlog.web.app as the main domain now
I was going to delete bodymetrics.web.app but I think I'll keep it for
now in case I want to use it to test out some other ideas for this as a
sort of staging environment.

This commit will hopefully provide a reference to how to change the
code to deploy to that environment as I always find this confusing!